+++ apps/web/CHANGELOG.md
@@ -1,5 +1,12 @@
 # web
 
+## 1.0.6
+
+### Patch Changes
+
+- Updated dependencies
+  - @swiss-digital-assets-institute/ui@1.0.4
+
 ## 1.0.5
 
 ### Patch Changes